Uploading some exercises of JavaFX

+++ García_Daniel_Jardinería/Entidad/Producto.cs
@@ -16,7 +16,7 @@ namespace Entidad
         public string Dimensiones { get; set; }
         public string Proveedor { get; set; }
         public string Descripcion { get; set; }
-        public short CantidadEnStock { get; set; }
+        public int CantidadEnStock { get; set; }
         public double PrecioVenta { get; set; }
         public double PrecioProveedor { get; set; }
